Update install script to use Git submodules where possible.
This commit elliminates the need to use the install script. A simple:

    $ git submodule update --init --recursive

would install most important components.

URxvt has also been updated to use the new ~/.urxvt/ext path.
The optional extension vim-scrollback was never any good so dump it.

It is still recommended to use the install script when GitHub or
rxvt-unicode would be used on the target machine.