Scripts: Add 'silent' to supress output of programs.
This script is useful when another program allows customisations of an
external command, but no way to supress its output. For example, several
Vim plug-ins can support a custom location for the ctags command,
however they allow no control to supress the output. By using
'silent ctags' as the command override that would be possible.