Vim: +FindClosest, like 'f' but looks in both directions.
I am fed up with having to decide if a character on the current line if
before or after the cursor. When I want to move to the '=' character,
it's usually the only one on the current line.

This function solves this for me. The mappings of ';' and ',' are also
kept consistent so it doesn't matter if the character was matched before
or after the cursor position.

W00t for being lazy.