Zsh: ^Z to toggle between an application and the shell.
This is much better than using tmux windows which are always two key
strokes away. ^Z in Vim, for example, will send the job to the
background returning control to the shell. The previous command-line
buffer will still be available. Upon using ^Z again, Vim is resumed.

The prompt will reflect when there are suspended jobs.