Scripts: `apt`, a thin wrapper around `aptdcon`.
The default `apt-get` sucks big time, especially if multiple processes
are trying to update the database. It requires 'root' privileges for
simple operations such as updating the packages cache.

`aptdcon` on the other hand is not very friendly on the command-line,
but extremely easy to use and only asks for 'root' when making changes
to the system, e.g., installing a new package.

`apt` is shorter to type, too.